Initial Commit
An app I wrote a long time ago following along with a Treehouse
tutorial.  The App varies in that instead of using a static array of
"fun facts", I wrote a quick and dirty AsyncTask to fetch facts from the
Cat Facts API.  I still have a static array of about 20 facts in case
network is unavailable or there is an error when trying to retrieve a
fact.